Fix interpreter exposure
Currently interpreter is being exposed as global variable inside
`DOMContentLoaded` event. This is not desirable, because other
JavaScript's are being loaded and executed before this event is even
called. Modify the loader so that the interpreter is created and
assigned to global variable immediately instead. This allows us to do
fancy tricks, so such overriding `print` method, which is being
demonstrated in the examples.